Grouping of symbolic variables in the same file and support to FixedBytes
